Elixir v1.4.4 Behaviour

This module has been deprecated.

Instead of defcallback/1 and defmacrocallback/1, the @callback and @macrocallback module attributes can be used (respectively). See the documentation for Module for more information on these attributes.

Instead of MyModule.__behaviour__(:callbacks), MyModule.behaviour_info(:callbacks) can be used.

Functions

defcallback(spec) (macro)

Defines a function callback according to the given type specification.

defmacrocallback(spec) (macro)

Defines a macro callback according to the given type specification.