Elixir v1.4.4 Kernel.ParallelRequire

A module responsible for requiring files in parallel.

Functions

files(files, callbacks \\ [])

Requires the given files.

A callback that will be invoked with each file, or a keyword list of callbacks can be provided:

:each_file - invoked with each file

:each_module - invoked with file, module name, and binary code

Returns the modules generated by each required file.