Mix v1.4.4 Mix.Shell behaviour

Defines Mix.Shell contract.

Functions

cmd(command, options \\ [], callback)

An implementation of the command callback that is shared across different shells.

printable_app_name()

Returns the printable app name.

This function returns the current application name, but only if the application name should be printed.

Calling this function automatically toggles its value to false until the current project is re-entered. The goal is to avoid printing the application name multiple times.

Callbacks

cmd(command)

cmd(command :: String.t) :: integer

Executes the given command and returns its exit status.

cmd(command, options)

cmd(command :: String.t, options :: Keyword.t) :: integer

Executes the given command and returns its exit status.

Options

:print_app - when false, does not print the app name when the command outputs something

:stderr_to_stdout - when false, does not redirect stderr to stdout

:quiet - when true, do not print the command output

:env - environment options to the executed command

error(message)

error(message :: IO.ANSI.ansidata) :: any

Prints the given error to the shell.

info(message)

info(message :: IO.ANSI.ansidata) :: any

Prints the given message to the shell.

print_app()

print_app() :: any

Prints the current application to the shell if it was not printed yet.

prompt(message)

prompt(message :: String.t) :: String.t

Prompts the user for input.

yes?(message)

yes?(message :: String.t) :: boolean

Prompts the user for confirmation.