Mix v1.4.4 Mix.Tasks.Compile.Yecc

Compiles Yecc source files.

When this task runs, it will check the modification time of every file, and if it has changed, the file will be compiled. Files will be compiled in the same source directory with a .erl extension. You can force compilation regardless of modification times by passing the --force option.

Command line options

--force - forces compilation regardless of modification times

Configuration

:erlc_paths - directories to find source files. Defaults to ["src"].

:yecc_options - compilation options that apply to Yecc’s compiler.

For a complete list of options, see :yecc.file/1. Note that the :report, :return_errors, and :return_warnings options are overridden by this compiler, thus setting them has no effect.

Functions

clean()

Cleans up compilation artifacts.

manifests()

Returns Yecc manifests.

run(args)

run(OptionParser.argv) :: :ok | :noop

Runs this task.