Mix v1.4.4 Mix.Tasks.Deps.Loadpaths

Checks and loads all dependencies along the way.

If there is an invalid dependency, its status is printed before aborting.

Although this task does not show up in mix help, it is part of Mix public API and can be depended on.

Command line options

--no-deps-check - does not check or compile deps, only load available ones

--no-compile - does not compile dependencies

Functions

run(args)

run(OptionParser.argv) :: :ok

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.