Mix v1.4.4 Mix.Tasks.Escript.Install

Installs an escript locally.

If no argument is supplied but there is an escript in the project’s root directory (created with mix escript.build), then the escript will be installed locally. For example:

mix do escript.build, escript.install

If an argument is provided, it should be a local path or a URL to a prebuilt escript, a git repository, a github repository, or a hex package.

mix escript.install escript
mix escript.install path/to/escript
mix escript.install https://example.com/my_escript
mix escript.install git https://path/to/git/repo
mix escript.install git https://path/to/git/repo branch git_branch
mix escript.install git https://path/to/git/repo tag git_tag
mix escript.install git https://path/to/git/repo ref git_ref
mix escript.install github user/project
mix escript.install github user/project branch git_branch
mix escript.install github user/project tag git_tag
mix escript.install github user/project ref git_ref
mix escript.install hex hex_package
mix escript.install hex hex_package 1.2.3

After installation, the escript can be invoked as

~/.mix/escripts/foo

For convenience, consider adding ~/.mix/escripts directory to your PATH environment variable. For more information, check the wikipedia article on PATH: https://en.wikipedia.org/wiki/PATH_(variable)

Command line options

--sha512 - checks the escript matches the given SHA-512 checksum. Only applies to installations via URL or local path.

--force - forces installation without a shell prompt; primarily intended for automation in build systems like make

--submodules - fetches repository submodules before building escript from git or github

--app - specifies a custom app name to be used for building the escript from git, github, or hex

Functions

build(mixfile)

For installs involving a fetch, this will be executed as the in_package.

Callback implementation for c:Mix.Local.Installer.build/1.

check_install_spec(, )

Checks that the install_spec and opts are supported by the respective module.

Callback implementation for c:Mix.Local.Installer.check_install_spec/2.

find_previous_versions(src, dst)

Returns a list of already installed version of the same archive or escript.

Callback implementation for c:Mix.Local.Installer.find_previous_versions/2.

install(dst, binary, previous)

The installation itself.

Callback implementation for c:Mix.Local.Installer.install/3.

run(argv)

run(OptionParser.argv) :: boolean

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.