Mix v1.4.4 Mix.Tasks.Help

Lists all tasks or prints the documentation for a given task.

Arguments

mix help                  - prints all tasks and their shortdoc
mix help TASK             - prints full docs for the given task
mix help --search PATTERN - prints all tasks that contain PATTERN in the name
mix help --names          - prints all task names and aliases
                            (useful for autocompleting)

Colors

When possible, mix help is going to use coloring for formatting guides. The formatting can be customized by configuring the Mix application either inside your project (in config/config.exs) or by using the local config (in ~/.mix/config.exs).

For example, to disable color, one may use the configuration:

[mix: [colors: [enabled: false]]]

The available color options are:

:enabled - shows ANSI formatting (defaults to IO.ANSI.enabled?)

:doc_code - the attributes for code blocks (cyan, bright)

:doc_inline_code - inline code (cyan)

:doc_headings - h1 and h2 (yellow, bright)

:doc_title - the overall heading for the output (reverse, yellow, bright)

:doc_bold - (bright)

:doc_underline - (underline)

Functions

run(argv)

run(OptionParser.argv) :: :ok

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.