Mix v1.4.4 Mix.Tasks.Local.PublicKeys

Public keys are used by Mix to install packages like Rebar and Hex.

Mix by default ships with a public key but new ones can be added on demand.

To list all available keys:

$ mix local.public_keys

To list all available keys showing the keys themselves:

$ mix local.public_keys --detailed

To add a new key:

$ mix local.public_keys local/path/to/key

Be careful when adding new keys. Only add keys from sources you trust.

Public keys are by default stored in your MIX_HOME under the public_keys directory.

Command line options

--force - forces installation without a shell prompt; primarily intended for automation in build systems like make

Functions

run(argv)

run(OptionParser.argv) :: true

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.